+++ HISTORY.md
@@ -2,12 +2,15 @@
 
 Note that the Racing+ Mod version will almost always match the client version. This means that the version number of the mod may increase, but have no actual new in-game changes. All gameplay related changes will be listed below.
 
+* *0.2.28* - Unreleased
+  * Changed the double Forsaken room to have two Dark Uriels.
+  * All Globins will permanently die upon the 5th regeneration to prevent Epic Fetus softlocks.
+  * Knights, Selfless Knights, Floating Knights, Bone Knights, Eyes, Bloodshot Eyes, Wizoobs, and Red Ghosts no longer have invulernability frames after spawning.
 * *0.2.27* - February 11th
   * Fixed a room with boils that can cause a softlock with no bombs or keys.
   * Fixed the double trouble room that softlocks if you have no bombs.
   * Added a bunch of splitting enemies to the fast-clear exception list.
   * Pedestal replacement no longer applies to shops. (It causes some weird behavior.)
-  * Changed the double Forsaken room to have two Mask of Infamys.
   * If you arrive at the Void floor for any reason, you will be automatically given the Polaroid and sent to Womb 1.
 * *0.2.20* - February 8th
   * Renamed the Jud6s mod to the Racing+ mod. This should still be considered alpha software. It contains all of the room changes from the Afterbirth version of the Jud6s mod, as well as things that were not in the Afterbirth version, like Wrath of the Lamb style room clear. The full list of changes are listed on [the main changes page](https://github.com/Zamiell/isaac-racing-mod/blob/master/CHANGES.md).